Refactor hydration and context handling. Clean up unused functions.
